Refactor: 프로젝트 패키지 구조 변경
- Group 내부 패키지 구조 변경
