+++ wekiki/src/main/java/com/kakao/techcampus/wekiki/comment/CommentResponse.java
@@ -1,6 +1,6 @@
 package com.kakao.techcampus.wekiki.comment;
 
-import com.kakao.techcampus.wekiki.group.member.GroupMember;
+import com.kakao.techcampus.wekiki.group.domain.member.GroupMember;
 import com.kakao.techcampus.wekiki.post.Post;
 import lombok.Getter;
 import lombok.Setter;
