+++ wekiki/src/main/java/com/kakao/techcampus/wekiki/group/domain/member/ActiveGroupMember.java
@@ -1,6 +1,6 @@
-package com.kakao.techcampus.wekiki.group.member;
+package com.kakao.techcampus.wekiki.group.domain.member;
 
-import com.kakao.techcampus.wekiki.group.Group;
+import com.kakao.techcampus.wekiki.group.domain.Group;
 import com.kakao.techcampus.wekiki.member.Member;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
