+++ wekiki/src/main/java/com/kakao/techcampus/wekiki/group/dto/requestDTO/JoinGroupRequestDTO.java
@@ -1,4 +1,4 @@
-package com.kakao.techcampus.wekiki.group.groupDTO.requestDTO;
+package com.kakao.techcampus.wekiki.group.dto.requestDTO;
 
 public record JoinGroupRequestDTO(
         String nickName
