+++ wekiki/src/main/java/com/kakao/techcampus/wekiki/group/dto/responseDTO/SearchOfficialGroupResponseDTO.java
@@ -1,6 +1,6 @@
-package com.kakao.techcampus.wekiki.group.groupDTO.responseDTO;
+package com.kakao.techcampus.wekiki.group.dto.responseDTO;
 
-import com.kakao.techcampus.wekiki.group.officialGroup.OfficialGroup;
+import com.kakao.techcampus.wekiki.group.domain.OfficialGroup;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.data.domain.Page;
