+++ wekiki/src/main/java/com/kakao/techcampus/wekiki/history/History.java
@@ -1,7 +1,6 @@
 package com.kakao.techcampus.wekiki.history;
 
-import com.kakao.techcampus.wekiki.group.member.GroupMember;
-import com.kakao.techcampus.wekiki.group.member.InactiveGroupMember;
+import com.kakao.techcampus.wekiki.group.domain.member.GroupMember;
 import com.kakao.techcampus.wekiki.post.Post;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
