+++ wekiki/src/main/java/com/kakao/techcampus/wekiki/history/HistoryJPARepository.java
@@ -1,6 +1,6 @@
 package com.kakao.techcampus.wekiki.history;
 
-import com.kakao.techcampus.wekiki.group.member.GroupMember;
+import com.kakao.techcampus.wekiki.group.domain.member.GroupMember;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
