+++ wekiki/src/main/java/com/kakao/techcampus/wekiki/member/MemberResponse.java
@@ -1,8 +1,8 @@
 package com.kakao.techcampus.wekiki.member;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.kakao.techcampus.wekiki.group.Group;
-import com.kakao.techcampus.wekiki.group.member.GroupMember;
+import com.kakao.techcampus.wekiki.group.domain.Group;
+import com.kakao.techcampus.wekiki.group.domain.member.GroupMember;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
