+++ wekiki/src/main/java/com/kakao/techcampus/wekiki/post/Post.java
@@ -1,7 +1,7 @@
 package com.kakao.techcampus.wekiki.post;
 
 import com.kakao.techcampus.wekiki.comment.Comment;
-import com.kakao.techcampus.wekiki.group.member.GroupMember;
+import com.kakao.techcampus.wekiki.group.domain.member.GroupMember;
 import com.kakao.techcampus.wekiki.history.History;
 import com.kakao.techcampus.wekiki.page.PageInfo;
 import jakarta.persistence.*;
