+++ wekiki/src/main/java/com/kakao/techcampus/wekiki/post/PostResponse.java
@@ -1,8 +1,7 @@
 package com.kakao.techcampus.wekiki.post;
 
-import com.kakao.techcampus.wekiki.group.member.GroupMember;
+import com.kakao.techcampus.wekiki.group.domain.member.GroupMember;
 import com.kakao.techcampus.wekiki.history.History;
-import com.kakao.techcampus.wekiki.page.PageInfo;
 import com.kakao.techcampus.wekiki.report.Report;
 import lombok.Getter;
 import lombok.Setter;
