+++ wekiki/src/main/java/com/kakao/techcampus/wekiki/report/Report.java
@@ -1,8 +1,7 @@
 package com.kakao.techcampus.wekiki.report;
 
-import com.kakao.techcampus.wekiki.group.member.GroupMember;
+import com.kakao.techcampus.wekiki.group.domain.member.GroupMember;
 import com.kakao.techcampus.wekiki.history.History;
-import com.kakao.techcampus.wekiki.post.Post;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
